refactor: move to schema dir
